+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## 1.7.0 (Released 2023-05-xx)
 
 - Updated to .NET 7.
+- A cancel button is now shown when in quick edit "mode".
 - A better error message is shown when loading an "EXT" vocals file fails.
 - Added a new validation check for hammer-ons/pull-offs that are on the same fret as the previous note on the same string.
 - Fixed false positive in validation for anchor position in relation to chord fingering with chords that use the thumb.